% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment-module.R
\name{comment_card}
\alias{comment_card}
\title{Create a comment card UI element}
\usage{
comment_card(proposal_id, intent, justification)
}
\arguments{
\item{proposal_id}{Unique ID for this proposal}

\item{intent}{Short description of the change}

\item{justification}{Detailed justification for the change}
}
\value{
Shiny tag
}
\description{
Create a comment card UI element
}
