% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment-module.R
\name{comment_mod_server}
\alias{comment_mod_server}
\title{Comment Pane Module Server}
\usage{
comment_mod_server(id, pending_edits, reviewing_started)
}
\arguments{
\item{id}{Module ID}

\item{pending_edits}{Reactive value containing list of pending edits}

\item{reviewing_started}{Reactive value indicating if review has started}
}
\value{
List of module functions
}
\description{
Comment Pane Module Server
}
