% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment-module.R
\name{comment_mod_ui}
\alias{comment_mod_ui}
\title{Comment Pane Module UI}
\usage{
comment_mod_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
Shiny UI element
}
\description{
Comment Pane Module UI
}
