% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_diffs.R
\name{diff_rmd}
\alias{diff_rmd}
\alias{diff_rmd_css}
\title{Render the differences between two rmarkdown files}
\usage{
diff_rmd(current_file, reference_file = "HEAD", show = "raw",
  output_format = "html_document", keep_intermediate = FALSE,
  quiet = TRUE, css = diff_rmd_css())

diff_rmd_css()
}
\arguments{
\item{current_file}{string: path to file after changes}

\item{reference_file}{string: path to file before changes}

\item{show}{string: \code{"raw"} (show differences in the raw rmarkdown file) or \code{"rendered"} (show differences in the rendered output). Note that the \code{"rendered"} output is unlikely to work well if a change has been made inside a code block in the rmarkdown document. The primary use for this package is anticipated to be the \code{"raw"} output}

\item{output_format}{string: format of the output file (currently only \code{"html_document"})}

\item{keep_intermediate}{logical: keep the intermediate rmarkdown file?}

\item{quiet}{logical: if \code{TRUE}, suppress pandoc output (Only applicable if \code{show="rendered"})}

\item{css}{character vector: css specification to apply to changed sections. Defaults to \code{diff_rmd_css()}; specify \code{NULL} to not include a \code{<style>} section in the output}
}
\value{
A list containing one or more elements \code{rendered} (the path to the rendered diff file, if \code{show="rendered"}), \code{intermediate} (the path to the intermediate file, if \code{keep_intermediate = TRUE}), and \code{raw} (if \code{show="raw"})

The path to the rendered file showing the differences
}
\description{
Render the differences between two rmarkdown files
}
\examples{
\dontrun{
  result <- diff_rmd(my_current_file, my_reference_file)
  browseURL(result)
}

}
