% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_rmd_addin.R
\name{diff_rmd_addin}
\alias{diff_rmd_addin}
\title{Rstudio addin to render the differences between two rmarkdown files}
\usage{
diff_rmd_addin()
}
\value{
Viewable html of the diff
}
\description{
This addin is a small wrapper for \code{diff_rmd}
}
