% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor-module.R
\name{editor_mod_server}
\alias{editor_mod_server}
\title{Editor Module Server}
\usage{
editor_mod_server(id, file_content, editable_region, pending_edits)
}
\arguments{
\item{id}{Module ID}

\item{file_content}{Reactive value containing file lines}

\item{editable_region}{Reactive value containing list(start, end)}

\item{pending_edits}{Reactive value containing list of pending edits}
}
\value{
List of module functions
}
\description{
Editor Module Server
}
