% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enable_html_annotation.R
\name{enable_html_annotation}
\alias{enable_html_annotation}
\title{Enable html annotations using the Hypothes.is web annontation client
(https://web.hypothes.is/)}
\usage{
enable_html_annotation()
}
\value{
Hypothes.is web client javascript text snippet
}
\description{
Enable html annotations using the Hypothes.is web annontation client
(https://web.hypothes.is/)
}
