% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_from_sas_to_r.R
\name{folder_from_sas_to_r}
\alias{folder_from_sas_to_r}
\title{folder_from_sas_to_r}
\usage{
folder_from_sas_to_r(path_folder, format_file)
}
\arguments{
\item{path_folder}{The data folder path.}

\item{format_file}{The format file.}
}
\value{
A list of the datamanaged files.
}
\description{
Import the whole folder and apply format file to it.
}
\examples{
\dontrun{
path_folder <- "folder_data/"
format_file <- "formats.sas7bdat"

res <- folder_from_sas_to_r(path_folder = path_folder,
format_file = format_file)
}

}
