% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-propose-edit.R
\name{make_propose_edit_impl}
\alias{make_propose_edit_impl}
\title{Factory to create the propose_edit implementation}
\usage{
make_propose_edit_impl()
}
\description{
Factory to create the propose_edit implementation
}
\keyword{internal}
