% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewer.R
\name{mypackageDependencies}
\alias{mypackageDependencies}
\title{mypackageDependencies
load Javascript and css assets required}
\usage{
mypackageDependencies()
}
\value{
An object that can be included in a list of dependencies passed to
attachDependencies.
}
\description{
mypackageDependencies
load Javascript and css assets required
}
\keyword{internal}
