% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewer.R
\name{pandoc}
\alias{pandoc}
\title{Call pandoc}
\usage{
pandoc(file = NULL, from, to, input = NULL)
}
\arguments{
\item{file}{length one character vector  path to input file}

\item{from}{length one character vector format to convert from}

\item{to}{length one character vector format to convert to}

\item{input}{character vector to be pased as stdin to pandoc}
}
\value{
length one character vector of converted document
}
\description{
Call pandoc
}
\keyword{internal}
