% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_folder_sas.R
\name{read_folder_sas}
\alias{read_folder_sas}
\title{read_folder_sas}
\usage{
read_folder_sas(path_folder, format_file = "formats.sas7bdat")
}
\arguments{
\item{path_folder}{The path folder (only .sas7bat files will be imported).}

\item{format_file}{The format file to excluded.}
}
\value{
A list of dataframe of the tables included inside the file.
}
\description{
A function to read all sas tables from a folder.
}
\examples{
\dontrun{
path_data_folder <- "folder_data/"
res <- read_sas(format_file = path_data_folder)
}

}
