% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_format.R
\name{read_format}
\alias{read_format}
\title{read_format}
\usage{
read_format(format_file)
}
\arguments{
\item{format_file}{The path to format file}
}
\value{
The sas format table
}
\description{
Function to read a SAS format file
}
\examples{
\dontrun{
path_format_file <- "formats.sas7bdat"
res <- read_format(format_file = path_format_file)
}

}
