% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sas.R
\name{read_sas}
\alias{read_sas}
\title{read_sas}
\usage{
read_sas(file)
}
\arguments{
\item{file}{The sas database file}
}
\value{
A dataframe with the data
}
\description{
A function to import sas data with the attributes.
}
\examples{
\dontrun{
path_data_file <- "data.sas7bdat"
res <- read_sas(format_file = path_data_file)
}

}
