% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_diffs.R
\name{diff_rmd}
\alias{diff_rmd}
\title{Render the differences between two rmarkdown files}
\usage{
diff_rmd(current_file, reference_file = "HEAD", output_format = "html",
  escape_code_chunks = FALSE)
}
\arguments{
\item{current_file}{string: path to file after changes}

\item{reference_file}{string: path to file before changes}

\item{output_format}{string: format of the output file (currently only \code{"html"})}

\item{escape_code_chunks}{logical: escape three-backtick code chunks?}
}
\value{
The path to the rendered file showing the differences
}
\description{
Render the differences between two rmarkdown files
}
\examples{
\dontrun{
  result <- diff_rmd(my_current_file, my_reference_file)
  browseURL(result)
}

}
