% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_diffs.R
\name{diff_rmd}
\alias{diff_rmd}
\title{Render the differences between two rmarkdown files}
\usage{
diff_rmd(current_file, reference_file = "HEAD",
  output_format = "html_document", keep_intermediate = FALSE,
  quiet = TRUE)
}
\arguments{
\item{current_file}{string: path to file after changes}

\item{reference_file}{string: path to file before changes}

\item{output_format}{string: format of the output file (currently only \code{"html"})}

\item{keep_intermediate}{logical: keep the intermediate rmarkdown file (which contains the marked-up differences)}

\item{quiet}{logical: if \code{TRUE}, suppress pandoc output during \code{rmarkdown::render} call}
}
\value{
A list containing elements \code{rendered} (the path to the rendered diff file) and (if \code{keep_intermediate = TRUE}) \code{intermediate} (the path to the intermediate file)
}
\description{
Render the differences between two rmarkdown files
}
\examples{
\dontrun{
  result <- diff_rmd(my_current_file, my_reference_file)
  browseURL(result$rendered)
}

}
