% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewer.R
\name{pandoc}
\alias{pandoc}
\title{Call pandoc}
\usage{
pandoc(input, from, to)
}
\arguments{
\item{input}{length one character vector  path to input file}

\item{from}{length one character vector format to convert from}

\item{to}{length one character vector format to convert to}
}
\value{
length one character vector of converted document
}
\description{
Call pandoc
}
\examples{
pandoc("test.docx","docx", "html")
}
\keyword{internal}
