% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewer-package.R
\docType{package}
\name{reviewer-package}
\alias{reviewer}
\alias{reviewer-package}
\title{reviewer: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
