% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewer-package.R
\docType{package}
\name{reviewer-package}
\alias{reviewer}
\alias{reviewer-package}
\title{reviewer: Improving the Track Changes and Reviewing Experience in R Markdown}
\description{
Provides functionality to compare two versions of an rmarkdown document and display their differences in a nicely-formatted manner, along with an RStudio addin that adds the required JavaScript code to an rmarkdown document, so that when rendered to HTML it can be annotated using the Hypothes.is service.
}
\author{
\strong{Maintainer}: Amy Stringer \email{amy.stringer@me.com}

Authors:
\itemize{
  \item Ben Raymond
  \item Mark Dulhunty
  \item Laura de Jong
}

}
\keyword{internal}
