% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewer.R
\name{reviewer}
\alias{reviewer}
\title{Launch addin to review changes
Launches and addin to review, accept or rect changes in a
Microsoft Word .docx file and then convert the reult to a markdown file}
\usage{
reviewer(wordDoc = NULL)
}
\arguments{
\item{wordDoc}{character path to Word .docx file}
}
\value{
character path to markdown file
}
\description{
Launch addin to review changes
Launches and addin to review, accept or rect changes in a
Microsoft Word .docx file and then convert the reult to a markdown file
}
\examples{
reviewer()
}
